Add the negative scenario code as well as their  corresponding test code.  Add jacoco too

+++ src/main/java/com/Api/service/UserService.java
@@ -3,13 +3,16 @@ package com.Api.service;
 import com.Api.Dao.UserRepositories;
 import com.Api.Entity.User;
 
+import com.Api.Exceptions.ServiceException;
 import org.hibernate.annotations.ColumnDefault;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
 import java.util.List;
 
 @Service
@@ -18,13 +21,23 @@ public class UserService {
     UserRepositories userRepositories;
 
     public List<User> findAll(){
-        return userRepositories.findAll();
+
+
+            List<User>  userList = userRepositories.findAll();
+            if(userList == null){
+                return Collections.emptyList();
+            }
+            return userList;
+
 
     }
 
     @Cacheable(value ="user",key = "#id",condition = "#id != null")
     public User findById(Integer id){
-        return  userRepositories.findById(id).orElseGet(()-> new User());
+        return  userRepositories.findById(id).orElseGet(()-> {
+            User user  = null;
+            return  user;
+        });
     }
 
 
@@ -34,6 +47,9 @@ public class UserService {
     }
     @CacheEvict(value = "user", key="#id")
     public void deleteById(Integer id){
+         if(!userRepositories.existsById(id)){
+             throw new ServiceException("User with ID " + id + " not found");
+         }
          userRepositories.deleteById(id);
     }
     @CachePut(value = "user", key = "#user.id")
@@ -48,7 +64,13 @@ public class UserService {
 
     }
     public User findByEmail(String email){
-        return userRepositories.getUserByEmail(email).orElseGet(()->new User());
+        return userRepositories.getUserByEmail(email).orElseGet(()->{
+            User user = null;
+            return user;
+        });
+    }
+    public User throwError() {
+        throw new NullPointerException("Testing aspect wrap");
     }
